light-oauth db script change, change the audit logid data type

+++ light-oauth2/sqlserver/create_sqlserver.sql
@@ -82,7 +82,7 @@ CREATE TABLE refresh_token (
 );
 
 create table audit_log (
-  log_id INT, -- system milliseonds from 1970.
+  log_id BigInt, -- system milliseonds from 1970.
   service_id VARCHAR(32) NOT NULL,
   endpoint VARCHAR(256) NOT NULL,
   request_header VARCHAR(4096),
